requirements.txt
torch>=1.5.0
torchaudio>=0.5.0
numpy
scipy
scikit-learn
tqdm
pyyaml
flask
flask-restful
